Updating PR
Updating PR with code review changes
